# api/openapi/valori-v1.yaml
components:
  schemas:
    ApiError:
      description: |-
        The canonical error body, as a schema-bearing DTO.

        [`valori_engine::EngineError`] produces this shape at runtime but lives in
        a crate that does not depend on `utoipa`, so the schema is declared here —
        the translation layer §36 asks for, rather than a re-export of an internal
        type.
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Stable machine-readable code. Branch on this.
        error:
          description: Human-readable message. Do not parse.
          type: string
      required:
      - error
      - code
      type: object
...
    ClusterRoleResponse:
      description: '`GET /v1/cluster/role` response.'
      properties:
        current_leader:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/u64'
            description: The leader this node currently believes in, if any.
        node_id:
          $ref: '#/components/schemas/u64'
        role:
          description: '`leader` or `follower`. Both are healthy.'
          type: string
      required:
      - role
      - node_id
      type: object
...
    ErrorCode:
      description: |-
        Mirror of [`valori_engine::ErrorCode`] for schema generation.

        `tests/api_contract.rs` diffs the runtime enum against the committed YAML;
        this type exists so the generated document carries the same closed set.
      enum:
      - validation_error
      - unauthorized
      - forbidden
      - not_found
      - collection_not_found
      - record_not_found
      - dimension_mismatch
      - invalid_metric
      - invalid_index
      - index_build_failed
      - conflict
      - capacity_exceeded
      - not_leader
      - unavailable
      - not_implemented
      - internal_error
      type: string
...
    MemberView:
      properties:
        api_addr:
          type: string
        id:
          $ref: '#/components/schemas/u64'
        raft_addr:
          type: string
        voter:
          type: boolean
      required:
      - id
      - raft_addr
      - api_addr
      - voter
      type: object
...
    StatusView:
      properties:
        current_leader:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/u64'
        is_leader:
          type: boolean
        last_applied_index:
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        last_log_index:
          format: int64
          minimum: 0
          type:
          - integer
          - 'null'
        members:
          items:
            $ref: '#/components/schemas/MemberView'
          type: array
        node_id:
          $ref: '#/components/schemas/u64'
        term:
          format: int64
          minimum: 0
          type: integer
      required:
      - node_id
      - is_leader
      - term
      - members
      type: object
...
    u64:
      format: int64
      minimum: 0
      type: integer
  securitySchemes:
    BearerAuth:
      description: Project API key or the legacy VALORI_AUTH_TOKEN bearer token.
      scheme: bearer
      type: http
info:
  description: 'The Valori Data Plane REST API v1, emitted directly from `#[utoipa::path]` annotations on the registered axum handlers. Every public route the Rust routers register appears here and nothing else does — `scripts/verify-api-route-contract.py` proves that equality on every run of the contract gate, and fails on any drift in either direction. Administrative and operator-internal routes (key management, cluster membership, replication streams, `/metrics`) are deliberately excluded: they are served by the node but are not part of the SDK surface.'
  license:
    identifier: MIT OR Apache-2.0
    name: MIT OR Apache-2.0
  title: Valori Data Plane API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/cluster/role:
    get:
      description: Always 200 — both `leader` and `follower` are healthy. A load balancer can steer writes with this instead of following redirects on every request.
      operationId: get_cluster_role
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClusterRoleResponse'
          description: This node's Raft role
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
      security:
      - BearerAuth: []
      summary: Whether this node is the leader
      tags:
      - cluster
      x-required-scope: read_only
      x-sdk: true
  /v1/cluster/status:
    get:
      description: Reported from this node's own metrics. `current_leader` is this node's current belief, not a quorum-confirmed fact.
      operationId: get_cluster_status
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusView'
          description: Cluster status
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'Missing or invalid credentials. The auth middleware rejects the request before the handler runs; `attach_error_code` renders the rejection as a canonical `ApiError` with `code: "unauthorized"`.'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: 'The presented key authenticated but its scope does not satisfy this operation''s `x-required-scope`. Rendered as a canonical `ApiError` with `code: "forbidden"`.'
      security:
      - BearerAuth: []
      summary: Raft membership and replication position
      tags:
      - cluster
      x-required-scope: read_only
      x-sdk: true
...
servers:
- description: 'The node this document was generated by. Valori is self-hosted: point your client at your own node''s base URL.'
  url: /