# crates/valori-node/src/cluster_api.rs
// Copyright (c) 2025 Varshith Gudur. Dual-licensed under MIT OR Apache-2.0.
//! Cluster management API — Phase 2.6.
//!
//! HTTP endpoints over the [`ClusterHandle`]'s Raft handle, mounted next to
//! the data-plane router when the node boots in cluster mode:
//!
//! | Method | Path | What |
//! |---|---|---|
//! | GET  | `/v1/cluster/status` | leader, term, indexes, membership |
//! | GET  | `/v1/cluster/health` | 200 when this node sees a leader, 503 otherwise |
//! | POST | `/v1/cluster/add-node` | learner-then-voter membership change |
//! | POST | `/v1/cluster/remove-node` | voter removal |
//!
//! Membership changes are leader-only: a follower answers **403 with the
//! leader's API address** so operators (and scripts) can retry against the
//! right node. Every accepted change goes through Raft itself — the
//! membership entry is committed like any other, so the change is durable
//! and ordered with respect to data writes.
//!
//! Admin-action audit events (`NodeJoined`/`NodeLeft` in the chained log)
//! land in Phase 2.9; these endpoints are the place they'll be emitted.

#[cfg(feature = "utoipa")]
#[allow(unused_imports)]
use crate::openapi::ApiError;

use axum::extract::{Query, State};
use axum::http::StatusCode;
use axum::response::{IntoResponse, Response};
use axum::routing::{get, post};
use axum::{Json, Router};
use serde::{Deserialize, Serialize};
use std::collections::{BTreeMap, BTreeSet};
use std::sync::Arc;

use valori_consensus::types::{NodeId, Raft, ShardId, ValoriNode};
use valori_wire::{AdminEvent, LogEntry};

use crate::events::event_log::EventLogWriter;

/// Shared state for the cluster endpoints.
#[derive(Clone)]
pub struct ClusterApiState {
    /// Shard 0's raft — every endpoint except `read_index` (Phase S6) only
    /// ever needs this one, since membership/status/health are shard-
    /// agnostic under S1's symmetric placement (every node is a voter in
    /// every shard, so shard 0's membership view is identical to every
    /// other shard's).
    pub raft: Arc<Raft>,
    /// Every shard this node runs, for the shard-aware read-index endpoint.
    /// Always contains at least `ShardId(0)` (== `raft` above).
    pub shards: Arc<BTreeMap<ShardId, Raft>>,
    /// Where successful membership changes are recorded as `AdminEvent`s
    /// in the BLAKE3 chain (Phase 2.9). `None` = no audit log configured.
    pub audit: Option<Arc<std::sync::Mutex<EventLogWriter>>>,
}

pub fn cluster_router(
    raft: Arc<Raft>,
    shards: Arc<BTreeMap<ShardId, Raft>>,
    audit: Option<Arc<std::sync::Mutex<EventLogWriter>>>,
) -> Router {
    Router::new()
        .route("/v1/cluster/status", get(status))
        .route("/v1/cluster/health", get(health))
        .route("/v1/cluster/read-index", get(read_index))
        .route("/v1/cluster/role", get(role))
        .route("/v1/cluster/add-node", post(add_node))
        .route("/v1/cluster/remove-node", post(remove_node))
        .route("/v1/cluster/snapshot", post(trigger_snapshot))
        .with_state(ClusterApiState {
            raft,
            shards,
            audit,
        })
}

/// Record an admin action in the chained audit log. Failures are logged,
/// not surfaced: the membership change has already committed through Raft —
/// the source of truth — and a full audit disk must not unwind it.
fn record_admin(state: &ClusterApiState, event: AdminEvent) {
    if let Some(audit) = &state.audit {
        let mut writer = audit.lock().expect("audit log mutex poisoned");
        if let Err(e) = writer.append(&LogEntry::Admin(event.clone())) {
            tracing::error!("failed to audit admin action {}: {e}", event.describe());
        }
    }
}

// ── Status ────────────────────────────────────────────────────────────────────

#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
#[derive(Serialize)]
pub(crate) struct MemberView {
    id: NodeId,
    raft_addr: String,
    api_addr: String,
    voter: bool,
}

#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
#[derive(Serialize)]
pub(crate) struct StatusView {
    node_id: NodeId,
    current_leader: Option<NodeId>,
    is_leader: bool,
    term: u64,
    last_log_index: Option<u64>,
    last_applied_index: Option<u64>,
    members: Vec<MemberView>,
}

#[cfg_attr(feature = "utoipa", utoipa::path(
    get,
    path = "/v1/cluster/status",
    operation_id = "get_cluster_status",
    tag = "cluster",
    summary = "Raft membership and replication position",
    description = "Reported from this node's own metrics. `current_leader` is this node's current belief, not a quorum-confirmed fact.",
    security(("BearerAuth" = [])),
    responses(
        (status = 200, description = "Cluster status", body = StatusView),
    ),
))]
async fn status(State(state): State<ClusterApiState>) -> Json<StatusView> {
    let m = state.raft.metrics().borrow().clone();

    let voters: std::collections::BTreeSet<NodeId> =
        m.membership_config.membership().voter_ids().collect();

    let members = m
        .membership_config
        .nodes()
        .map(|(id, node)| MemberView {
            id: *id,
            raft_addr: node.raft_addr.clone(),
            api_addr: node.api_addr.clone(),
            voter: voters.contains(id),
        })
        .collect();

    Json(StatusView {
        node_id: m.id,
        current_leader: m.current_leader,
        is_leader: m.current_leader == Some(m.id),
        term: m.current_term,
        last_log_index: m.last_log_index,
        last_applied_index: m.last_applied.map(|l| l.index),
        members,
    })
}

// ── Read index (linearizable reads) ─────────────────────────────────────────────
//
// The read-index protocol's leader half. `get_read_log_id` confirms this node
// is still the leader (via a heartbeat round to a quorum) and returns the log
// index a read must observe to be linearizable. A follower calls this, then
// waits until its own applied index reaches `read_index` before serving —
// see `ensure_read_consistency` in cluster_server.rs.

#[derive(Deserialize, Default)]
struct ReadIndexQuery {
    /// Phase S6: which shard's read index to establish. Absent = shard 0,
    /// byte-identical to pre-S6 behavior for every caller that doesn't know
    /// about sharding yet.
    #[serde(default)]
    shard: u32,
}

async fn read_index(
    State(state): State<ClusterApiState>,
    Query(q): Query<ReadIndexQuery>,
) -> Response {
    let shard_id = ShardId(q.shard);
    let raft = match state.shards.get(&shard_id) {
        Some(r) => r,
        None => {
            return (
                StatusCode::NOT_FOUND,
                Json(serde_json::json!({ "error": format!("unknown shard {}", q.shard) })),
            )
                .into_response();
        }
    };

    match raft.get_read_log_id().await {
        Ok((read_log_id, _applied)) => (
            StatusCode::OK,
            Json(serde_json::json!({
                "read_index": read_log_id.map(|l| l.index).unwrap_or(0),
                "shard": shard_id.0,
            })),
        )
            .into_response(),
        Err(e) => {
            // Not the leader anymore, or no quorum — name the leader we know so
            // the caller can re-resolve and retry.
            let leader = raft.metrics().borrow().current_leader;
            (
                StatusCode::SERVICE_UNAVAILABLE,
                Json(serde_json::json!({
                    "error": "cannot establish read index (not leader or no quorum)",
                    "leader": leader,
                    "shard": shard_id.0,
                    "detail": format!("{e}"),
                })),
            )
                .into_response()
        }
    }
}

// ── Role (for load-balancer leader routing) ───────────────────────────────────
//
// A load balancer can route POST /records to the leader by health-checking
// this endpoint. On the leader it returns 200 + {"role":"leader"}; on a
// follower it returns 200 + {"role":"follower"}. Both are healthy — the
// distinction lets the LB steer writes without the SDK needing to follow
// redirects for every request.

/// `GET /v1/cluster/role` response.
#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
#[derive(Serialize)]
pub(crate) struct ClusterRoleResponse {
    /// `leader` or `follower`. Both are healthy.
    role: String,
    node_id: NodeId,
    /// The leader this node currently believes in, if any.
    current_leader: Option<NodeId>,
}

/// `GET /v1/cluster/health` response.
#[cfg_attr(feature = "utoipa", derive(utoipa::ToSchema))]
#[derive(Serialize)]
pub(crate) struct ClusterHealthResponse {
    /// `ok` when a leader is visible, `no-leader` otherwise.
    status: String,
    #[serde(skip_serializing_if = "Option::is_none")]
    leader: Option<NodeId>,
    /// Present only on the `no-leader` path.
    #[serde(skip_serializing_if = "Option::is_none")]
    detail: Option<String>,
}

#[cfg_attr(feature = "utoipa", utoipa::path(
    get,
    path = "/v1/cluster/role",
    operation_id = "get_cluster_role",
    tag = "cluster",
    summary = "Whether this node is the leader",
    description = "Always 200 — both `leader` and `follower` are healthy. A load balancer can steer writes with this instead of following redirects on every request.",
    security(("BearerAuth" = [])),
    responses(
        (status = 200, description = "This node's Raft role", body = ClusterRoleResponse),
    ),
))]
async fn role(State(state): State<ClusterApiState>) -> Json<ClusterRoleResponse> {
    let m = state.raft.metrics().borrow().clone();
    let is_leader = m.current_leader == Some(m.id);
    Json(ClusterRoleResponse {
        role: if is_leader { "leader" } else { "follower" }.to_string(),
        node_id: m.id,
        current_leader: m.current_leader,
    })
}

// ── Health ────────────────────────────────────────────────────────────────────

#[cfg_attr(feature = "utoipa", utoipa::path(
    get,
    path = "/v1/cluster/health",
    operation_id = "get_cluster_health",
    tag = "cluster",
    summary = "Whether this node sees an elected leader",
    description = "503 with `status: no-leader` during an election. Distinct from `GET /health`, which reports this node's own serving capacity.",
    security(("BearerAuth" = [])),
    responses(
        (status = 200, description = "A leader is elected", body = ClusterHealthResponse),
        (status = 503, description = "No leader currently visible from this node", body = ClusterHealthResponse),
    ),
))]
async fn health(State(state): State<ClusterApiState>) -> Response {
    let m = state.raft.metrics().borrow().clone();
    match m.current_leader {
        Some(leader) => (
            StatusCode::OK,
            Json(ClusterHealthResponse {
                status: "ok".to_string(),
                leader: Some(leader),
                detail: None,
            }),
        )
            .into_response(),
        None => (
            StatusCode::SERVICE_UNAVAILABLE,
            Json(ClusterHealthResponse {
                status: "no-leader".to_string(),
                leader: None,
                detail: Some("this node currently sees no elected leader".to_string()),
            }),
        )
            .into_response(),
    }
}

// ── Membership changes ────────────────────────────────────────────────────────

#[derive(Deserialize)]
struct AddNodeRequest {
    node_id: NodeId,
    raft_addr: String,
    #[serde(default)]
    api_addr: String,
}

#[derive(Deserialize)]
struct RemoveNodeRequest {
    node_id: NodeId,
}

/// Maps a leader-only failure to 403 + the leader's API address.
fn leadership_error<E>(raft_err: openraft::error::RaftError<NodeId, E>) -> Response
where
    E: std::fmt::Display,
{
    let body = match &raft_err {
        openraft::error::RaftError::APIError(e) => serde_json::json!({
            "error": "not-leader-or-rejected",
            "detail": e.to_string(),
        }),
        openraft::error::RaftError::Fatal(e) => serde_json::json!({
            "error": "raft-fatal",
            "detail": e.to_string(),
        }),
    };
    (StatusCode::FORBIDDEN, Json(body)).into_response()
}

async fn add_node(
    State(state): State<ClusterApiState>,
    Json(req): Json<AddNodeRequest>,
) -> Response {
    let node = ValoriNode {
        raft_addr: req.raft_addr,
        api_addr: req.api_addr,
    };

    // Step 1: add as learner — it catches up on the log (or receives a
    // snapshot, Phase 2.7) without affecting quorum.
    if let Err(e) = state
        .raft
        .add_learner(req.node_id, node.clone(), true)
        .await
    {
        return leadership_error(e);
    }

    // Step 2: promote to voter. The address is already registered by
    // add_learner, so the change is expressed as the new voter-id set.
    // `retain: false` — the node moves from learner to voter rather than
    // being duplicated in both sets.
    let mut voters: BTreeSet<NodeId> = state
        .raft
        .metrics()
        .borrow()
        .membership_config
        .membership()
        .voter_ids()
        .collect();
    voters.insert(req.node_id);

    match state.raft.change_membership(voters, false).await {
        Ok(resp) => {
            record_admin(
                &state,
                AdminEvent::NodeJoined {
                    node_id: req.node_id,
                    raft_addr: node.raft_addr.clone(),
                    api_addr: node.api_addr.clone(),
                    // All-zeros until RBAC lands (Phase 3): "no auth configured".
                    authorized_by: [0u8; 16],
                },
            );
            (
                StatusCode::OK,
                Json(serde_json::json!({
                    "status": "added",
                    "node_id": req.node_id,
                    "log_index": resp.log_id.index,
                })),
            )
                .into_response()
        }
        Err(e) => leadership_error(e),
    }
}

async fn remove_node(
    State(state): State<ClusterApiState>,
    Json(req): Json<RemoveNodeRequest>,
) -> Response {
    let remaining: BTreeSet<NodeId> = state
        .raft
        .metrics()
        .borrow()
        .membership_config
        .membership()
        .voter_ids()
        .filter(|id| *id != req.node_id)
        .collect();

    if remaining.is_empty() {
        return (
            StatusCode::UNPROCESSABLE_ENTITY,
            Json(serde_json::json!({
                "error": "cannot-remove-last-voter",
                "detail": "removing this node would leave the cluster with no voters",
            })),
        )
            .into_response();
    }

    // `retain: false`: the removed voter is dropped entirely, not demoted
    // to a lingering learner.
    match state.raft.change_membership(remaining, false).await {
        Ok(resp) => {
            record_admin(
                &state,
                AdminEvent::NodeLeft {
                    node_id: req.node_id,
                    authorized_by: [0u8; 16],
                },
            );
            (
                StatusCode::OK,
                Json(serde_json::json!({
                    "status": "removed",
                    "node_id": req.node_id,
                    "log_index": resp.log_id.index,
                })),
            )
                .into_response()
        }
        Err(e) => leadership_error(e),
    }
}

// ── Snapshot trigger ──────────────────────────────────────────────────────────
// Asks the local Raft core to build a snapshot immediately. This compacts the
// log up to last_applied, so a freshly-joined node will receive the snapshot
// via InstallSnapshot rather than replaying the full log. Useful for testing
// and for admins who want to force a checkpoint before adding a new replica.

async fn trigger_snapshot(State(state): State<ClusterApiState>) -> Response {
    match state.raft.trigger().snapshot().await {
        Ok(()) => (
            StatusCode::OK,
            Json(serde_json::json!({ "status": "triggered" })),
        )
            .into_response(),
        Err(e) => (
            StatusCode::INTERNAL_SERVER_ERROR,
            Json(serde_json::json!({
                "error": "snapshot trigger failed",
                "detail": format!("{e}"),
            })),
        )
            .into_response(),
    }
}